% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat.R
\name{cal_multiply_records}
\alias{cal_multiply_records}
\title{Calculate multiply transaction records VAT}
\usage{
cal_multiply_records(
  x,
  symbol = "symbol",
  amount = "amount",
  volume = "volume",
  tax_rate = 0.06,
  ...
)
}
\arguments{
\item{x}{A data frame containing the multiply transaction records}

\item{symbol}{The symbol column name of the transaction}

\item{amount}{The amount column name of the transaction}

\item{volume}{The volume column name of the transaction}

\item{tax_rate}{The tax rate}

\item{...}{The director column name of the transaction,can be null or length 1}
}
\value{
A data frame containing the transaction records with VAT calculated
}
\description{
This function calculates the VAT of a series of transaction records.Moving weighted average method is used to calculate the average holding price.
}
\examples{
data(records)
cal_multiply_records(x=records)
}
