% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vat.R
\name{calculate_financial_transfer_vat}
\alias{calculate_financial_transfer_vat}
\title{Calculate financial transfer VAT}
\usage{
calculate_financial_transfer_vat(dates, differences, vat_rate = 0.06)
}
\arguments{
\item{dates}{A vector of dates}

\item{differences}{A vector of differences}

\item{vat_rate}{The tax rate}
}
\value{
A data frame containing the VAT of financial transfer
}
\description{
This function calculates the VAT of financial transfer
}
\examples{
dates <- as.Date(c("2021-01-01", "2021-02-02", "2022-01-03", "2022-01-04", "2022-05-05"))
differences <- c(100, -50, 200, -150, -300)
calculate_financial_transfer_vat(dates, differences)
}
