% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{records}
\alias{records}
\title{A dataset containing 10 records of transactions.
 @name records
 @title sample data
 @format A data frame with 10 rows and 4 variables
\describe{
  \item{amount}{transaction amount.}
  \item{volume}{transaction volume.}
}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 5 columns.
}
\source{
A data frame with 10 rows and 4 variables
}
\usage{
records
}
\description{
A dataset containing 10 records of transactions.
 @name records
 @title sample data
 @format A data frame with 10 rows and 4 variables
\describe{
  \item{amount}{transaction amount.}
  \item{volume}{transaction volume.}
}
}
\examples{
data('records')
}
\keyword{datasets}
